+++ docs/Output/Dhtml/pickledFeatures.js
@@ -3084,7 +3084,7 @@ jsonPWrapper ({
   },
   "Configuration": {
     "SutName": "Pickles",
-    "SutVersion": "2.13.0",
-    "GeneratedOn": "10 Februar 2017 11:00:51"
+    "SutVersion": "2.14.0",
+    "GeneratedOn": "24 Februar 2017 18:09:34"
   }
 });