+++ src/Pickles/Pickles.Test/BaseFixture.cs
@@ -43,9 +43,12 @@ namespace PicklesDoc.Pickles.Test
                 if (this.container == null)
                 {
                     var builder = new ContainerBuilder();
+
+                    var configuration = new Configuration() { ExcludeTags = "exclude-tag" };
                     builder.RegisterAssemblyTypes(typeof(Runner).Assembly);
                     builder.Register<MockFileSystem>(_ => CreateMockFileSystem()).As<IFileSystem>().SingleInstance();
                     builder.RegisterModule<PicklesModule>();
+                    builder.RegisterInstance(configuration).As<IConfiguration>().SingleInstance();
                     this.container = builder.Build();
                 }
 
@@ -88,7 +91,7 @@ namespace PicklesDoc.Pickles.Test
             this.AddFakeFolderAndFiles("AcceptanceTest", new[] { "AdvancedFeature.feature", "LevelOne.feature" });
             this.AddFakeFolderAndFiles("EmptyFolderTests", new string[0]);
 
-            this.AddFakeFolderAndFiles("FeatureCrawlerTests", new[] { "index.md", "LevelOne.feature", "image.png" });
+            this.AddFakeFolderAndFiles("FeatureCrawlerTests", new[] { "index.md", "LevelOne.feature", "image.png", "LevelOneIgnoredFeature.feature" });
             this.AddFakeFolderAndFiles(@"FeatureCrawlerTests\SubLevelOne", new[] { "ignorethisfile.ignore", "LevelOneSublevelOne.feature", "LevelOneSublevelTwo.feature" });
             this.AddFakeFolderAndFiles(@"FeatureCrawlerTests\SubLevelOne\SubLevelTwo", new[] { "LevelOneSublevelOneSubLevelTwo.feature" });
             this.AddFakeFolderAndFiles(@"FeatureCrawlerTests\SubLevelOne\SubLevelTwo\IgnoreThisDirectory", new[] { "IgnoreThisFile.ignore" });