+++ src/Pickles/Pickles.Test/WhenParsingFeatureFiles.cs
@@ -18,10 +18,10 @@
 //  </copyright>
 //  --------------------------------------------------------------------------------------------------------------------
 
-using System;
 using System.Linq;
 
 using Autofac;
+using DocumentFormat.OpenXml.Bibliography;
 using NFluent;
 using NUnit.Framework;
 using PicklesDoc.Pickles.Extensions;
@@ -395,5 +395,133 @@ Feature: Test
         }
 
 
+
+        [Test]
+        public void Then_can_parse_and_ignore_feature_with_tag_in_configuration_ignore_tag()
+        {
+            var featureText =
+                @"# ignore this comment
+@feature-tag @exclude-tag
+Feature: Test
+    In order to do something
+    As a user
+    I want to run this scenario
+
+    @scenario-tag-1 @scenario-tag-2
+  Scenario: A scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens";
+
+            var parser = Container.Resolve<FeatureParser>();
+            var feature = parser.Parse(new StringReader(featureText));
+            Check.That(feature).IsNull();
+        }
+
+        [Test]
+        public void Then_can_parse_and_ignore_scenario_with_tag_in_configuration_ignore_tag()
+        {
+            var featureText =
+                @"# ignore this comment
+@feature-tag
+Feature: Test
+    In order to do something
+    As a user
+    I want to run this scenario
+
+    @scenario-tag-1 @scenario-tag-2
+  Scenario: A scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens
+
+    @scenario-tag-1 @scenario-tag-2 @exclude-tag
+  Scenario: B scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens
+
+    @scenario-tag-1 @scenario-tag-2
+  Scenario: C scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens";
+ 
+            var parser = Container.Resolve<FeatureParser>();
+            var feature = parser.Parse(new StringReader(featureText));
+
+            Check.That(feature.FeatureElements.Count).IsEqualTo(2);
+            Check.That(feature.FeatureElements.FirstOrDefault(fe => fe.Name == "A scenario")).IsNotNull();
+            Check.That(feature.FeatureElements.FirstOrDefault(fe => fe.Name == "B scenario")).IsNull();
+            Check.That(feature.FeatureElements.FirstOrDefault(fe => fe.Name == "C scenario")).IsNotNull();
+        }
+
+        [Test]
+        public void Then_can_parse_and_ignore_scenario_with_tag_in_configuration_ignore_tag_and_keep_feature()
+        {
+            var featureText =
+                @"# ignore this comment
+@feature-tag
+Feature: Test
+    In order to do something
+    As a user
+    I want to run this scenario
+
+    @scenario-tag-1 @scenario-tag-2 @Exclude-Tag
+  Scenario: A scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens
+
+    @scenario-tag-1 @scenario-tag-2 @exclude-tag
+  Scenario: B scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens";
+
+            var parser = Container.Resolve<FeatureParser>();
+            var feature = parser.Parse(new StringReader(featureText));
+
+            Check.That(feature).IsNotNull();
+            Check.That(feature.FeatureElements).IsEmpty();
+        }
+
+        [Test]
+        public void Then_can_parse_and_ignore_with_with_tag_without_sensitivity()
+        {
+
+            var featureText =
+                @"# ignore this comment
+@feature-tag
+Feature: Test
+    In order to do something
+    As a user
+    I want to run this scenario
+
+    @scenario-tag-1 @scenario-tag-2 @Exclude-Tag
+  Scenario: A scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens
+
+    @scenario-tag-1 @scenario-tag-2 @exclude-tag
+  Scenario: B scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens
+
+    @scenario-tag-1 @scenario-tag-2 @ExClUdE-tAg
+  Scenario: C scenario
+    Given some feature
+    When it runs
+    Then I should see that this thing happens";
+
+            var parser = Container.Resolve<FeatureParser>();
+            var feature = parser.Parse(new StringReader(featureText));
+
+            Check.That(feature).IsNotNull();
+            Check.That(feature.FeatureElements).IsEmpty();
+        }
     }
 }
+