+++ src/Pickles/Pickles/ConfigurationReporter.cs
@@ -35,6 +35,7 @@ namespace PicklesDoc.Pickles
             writeToLog($"Language                       : {configuration.Language}");
             writeToLog($"Incorporate Test Results?      : {(configuration.HasTestResults ? "Yes" : "No")}");
             writeToLog($"Include Experimental Features? : {(configuration.ShouldIncludeExperimentalFeatures ? "Yes" : "No")}");
+            writeToLog($"Exclude Tag                    : {configuration.ExcludeTags}");
 
             if (configuration.HasTestResults)
             {